# api/openapi.yaml
openapi: 3.0.3
info:
  title: Nabu API
  description: >
    Self-hosted task tracker for software development teams. This spec is the
    source of truth for the API contract: backend server types and the
    TypeScript client are both generated from it (docs/conventions/api-contract.md).
  version: 0.1.0

paths:
  /health:
    get:
      operationId: getHealth
      summary: Liveness check
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"

  /api/v1/auth/login:
    post:
      operationId: login
      summary: Exchange credentials for a session
      description: >
        On success, sets the access and refresh tokens as HTTP-only cookies and
        returns the authenticated user. Credentials are never echoed back.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated; session cookies set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/auth/refresh:
    post:
      operationId: refresh
      summary: Rotate the session using the refresh cookie
      description: >
        Reads the refresh cookie, rotates it, and sets fresh access and refresh
        cookies. No request body; the refresh token is never exposed to JS.
      responses:
        "204":
          description: Session rotated; new cookies set
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current session
      description: >
        Revokes the refresh-token family and clears the session cookies.
        Idempotent — always succeeds so the client can clear local state.
      responses:
        "204":
          description: Session revoked and cookies cleared

  /api/v1/users/me:
    get:
      operationId: getCurrentUser
      summary: The authenticated user
      description: >
        Deliberately outside /api/v1/auth so the refresh cookie's Path scope
        does not attach the long-lived refresh token to this per-page-load call
        (ADR-0003).
      responses:
        "200":
          description: The current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/projects:
    get:
      operationId: listProjects
      summary: List the projects in the caller's organization
      description: >
        Unpaginated: a self-hosted org has a handful of projects, and the list
        populates a selector rather than a scrollable view.
      responses:
        "200":
          description: Every project the caller's organization owns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/tasks:
    get:
      operationId: listTasks
      summary: List tasks, newest first, cursor-paginated
      parameters:
        - name: projectId
          in: query
          required: false
          description: Restrict to one project. Omit for every project in the org.
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: cursor
          in: query
          required: false
          description: Opaque cursor from a previous response's nextCursor.
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: Defaults to 50, capped at 100.
          schema:
            type: integer
      responses:
        "200":
          description: One page of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      operationId: createTask
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: The created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"

components:
  responses:
    ValidationError:
      description: Request was well-formed but failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or expired session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    UserProfile:
      type: object
      required: [id, email, displayName]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string

    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string

    TaskStatus:
      type: string
      enum: [todo, in_progress, done]

    Project:
      type: object
      required: [id, key, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
          description: Short human-facing identifier, unique per organization.
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProjectList:
      type: object
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        nextCursor:
          type: string
          nullable: true
          description: >
            Always null. Projects are a bounded per-org collection, so the
            endpoint is not paginated yet; the field is present so every list
            response carries the same envelope.

    Task:
      type: object
      required: [id, projectId, title, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        title:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateTaskRequest:
      type: object
      required: [title, projectId]
      properties:
        title:
          type: string
          maxLength: 500
        projectId:
          type: string
          format: uuid
          description: Must be a project in the caller's organization.

    TaskList:
      type: object
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        nextCursor:
          type: string
          nullable: true
          description: Opaque; pass back as the cursor query param. Null on the last page.

    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable identifier; clients switch on this.
        message:
          type: string
          description: Human-readable; wording may change, never parse it.

    Error:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"