# internal/project/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlcgen

import (
	"time"

	"github.com/google/uuid"
)

type Project struct {
	ID        uuid.UUID
	OrgID     uuid.UUID
	Key       string
	Name      string
	CreatedAt time.Time
	UpdatedAt time.Time
}